move get_evaluation_metrics to rasa.test
