+++ rasa/nlu/test.py
@@ -83,62 +83,6 @@ def log_evaluation_table(
     logger.info(f"Classification report: \n{report}")
 
 
-def get_evaluation_metrics(
-    targets: Iterable[Any],
-    predictions: Iterable[Any],
-    output_dict: bool = False,
-    exclude_label: Text = None,
-) -> Tuple[Union[Text, Dict[Text, Dict[Text, float]]], float, float, float]:
-    """Compute the f1, precision, accuracy and summary report from sklearn.
-
-    Args:
-        targets: target labels
-        predictions: predicted labels
-        output_dict: if True sklearn returns a summary report as dict, if False the
-          report is in string format
-        exclude_label: labels to exclude from evaluation
-
-    Returns: a report from sklearn, precision, f1, and accuracy values
-    """
-    from sklearn import metrics
-
-    targets = clean_labels(targets)
-    predictions = clean_labels(predictions)
-
-    labels = get_unique_labels(targets, exclude_label)
-    if not labels:
-        logger.warning("No labels to evaluate. Skip evaluation.")
-        return {}, 0.0, 0.0, 0.0
-
-    report = metrics.classification_report(
-        targets, predictions, labels=labels, output_dict=output_dict
-    )
-    precision = metrics.precision_score(
-        targets, predictions, labels=labels, average="weighted"
-    )
-    f1 = metrics.f1_score(targets, predictions, labels=labels, average="weighted")
-    accuracy = metrics.accuracy_score(targets, predictions)
-
-    return report, precision, f1, accuracy
-
-
-def get_unique_labels(
-    targets: Iterable[Text], exclude_label: Optional[Text]
-) -> List[Text]:
-    """Get unique labels. Exclude 'exclude_label' if specified.
-
-    Args:
-        targets: labels
-        exclude_label: label to exclude
-
-    Returns: unique list of labels
-    """
-    labels = set(targets)
-    if exclude_label and exclude_label in labels:
-        labels.remove(exclude_label)
-    return list(labels)
-
-
 def remove_empty_intent_examples(
     intent_results: List[IntentEvaluationResult],
 ) -> List[IntentEvaluationResult]:
@@ -184,17 +128,6 @@ def remove_empty_response_examples(
             filtered.append(r)
 
     return filtered
-
-
-def clean_labels(labels: Iterable[Text]) -> List[Text]:
-    """Remove `None` labels. sklearn metrics do not support them.
-
-    Args:
-        labels: list of labels
-
-    Returns: cleaned labels
-    """
-    return [label if label is not None else "" for label in labels]
 
 
 def drop_intents_below_freq(
@@ -445,6 +378,7 @@ def evaluate_response_selections(
     """
     import sklearn.metrics
     import sklearn.utils.multiclass
+    from rasa.test import get_evaluation_metrics
 
     # remove empty response targets
     num_examples = len(response_selection_results)
@@ -627,6 +561,7 @@ def evaluate_intents(
     """
     import sklearn.metrics
     import sklearn.utils.multiclass
+    from rasa.test import get_evaluation_metrics
 
     # remove empty intent targets
     num_examples = len(intent_results)
@@ -921,6 +856,7 @@ def evaluate_entities(
     """
     import sklearn.metrics
     import sklearn.utils.multiclass
+    from rasa.test import get_evaluation_metrics
 
     aligned_predictions = align_all_entity_predictions(entity_results, extractors)
     merged_targets = merge_labels(aligned_predictions)
@@ -1967,6 +1903,8 @@ def _compute_metrics(
 
     Returns: metrics
     """
+    from rasa.test import get_evaluation_metrics
+
     # compute fold metrics
     targets, predictions = _targets_predictions_from(
         results, target_key, prediction_key
@@ -1987,6 +1925,8 @@ def _compute_entity_metrics(
 
     Returns: entity metrics
     """
+    from rasa.test import get_evaluation_metrics
+
     entity_metric_results: EntityMetrics = defaultdict(lambda: defaultdict(list))
     extractors = get_entity_extractors(interpreter)
